refactor: extract leaderboard cluster from GameState into LeaderboardMixin (#1271)
