+++ custom_components/beatify/game/state_leaderboard.py
@@ -1,0 +1,131 @@
+"""Leaderboard subsystem for :class:`GameState`.
+
+Issue #1271 next-increment extraction: the leaderboard / ranking cluster
+(``get_leaderboard``, ``_store_previous_ranks``, ``get_final_leaderboard`` —
+originally Stories 5.5 / 5.6) is pulled out of the ``game/state.py``
+God-Object into this ``LeaderboardMixin``.
+
+The mixin is **behavior-preserving**: it carries the exact same methods that
+previously lived on ``GameState``. ``GameState`` inherits them, so its public
+API and every caller / test are unchanged.
+
+The mixin relies on a single attribute the host class owns and that lives on
+``self`` at runtime:
+
+* ``self.players`` — mapping of player name → ``PlayerSession``. Each session
+  exposes ``score``, ``name``, ``streak``, ``is_admin``, ``connected``,
+  ``previous_rank``, ``best_streak``, ``rounds_played`` and ``bets_won``.
+
+It carries no state of its own and imports nothing from ``state.py``, so the
+extraction introduces no cyclic imports.
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+
+class LeaderboardMixin:
+    """Leaderboard / ranking behavior for :class:`GameState`.
+
+    See module docstring for the host-class attributes this mixin reads.
+    """
+
+    def get_leaderboard(self) -> list[dict[str, Any]]:
+        """
+        Get leaderboard sorted by score (Story 5.5).
+
+        Returns:
+            List of player data with rank and movement info.
+            Note: is_current is set client-side based on playerName.
+
+        """
+        # Sort by score descending, then by name for tie-breaking display order
+        sorted_players = sorted(
+            self.players.values(),
+            key=lambda p: (-p.score, p.name),
+        )
+
+        leaderboard = []
+        current_rank = 0
+        previous_score = None
+
+        for i, player in enumerate(sorted_players):
+            # Handle ties (same score = same rank)
+            # Example: scores [100, 80, 80, 50] -> ranks [1, 2, 2, 4]
+            if player.score != previous_score:
+                current_rank = i + 1  # Rank jumps to position (skips tied ranks)
+            previous_score = player.score
+
+            # Calculate rank change (positive = moved up)
+            rank_change = 0
+            if player.previous_rank is not None:
+                rank_change = player.previous_rank - current_rank
+
+            entry = {
+                "rank": current_rank,
+                "name": player.name,
+                "score": player.score,
+                "streak": player.streak,
+                "is_admin": player.is_admin,
+                "rank_change": rank_change,
+                "connected": player.connected,
+            }
+            leaderboard.append(entry)
+
+        return leaderboard
+
+    def _store_previous_ranks(self) -> None:
+        """Store current ranks before scoring for rank change detection."""
+        sorted_players = sorted(
+            self.players.values(),
+            key=lambda p: (-p.score, p.name),
+        )
+
+        current_rank = 0
+        previous_score = None
+
+        for i, player in enumerate(sorted_players):
+            if player.score != previous_score:
+                current_rank = i + 1
+            previous_score = player.score
+            player.previous_rank = current_rank
+
+    def get_final_leaderboard(self) -> list[dict[str, Any]]:
+        """
+        Get final leaderboard with full player stats (Story 5.6).
+
+        Returns:
+            List of player data with rank and final stats.
+            Note: is_current is set client-side based on playerName.
+
+        """
+        # Sort by score descending, then by name for tie-breaking display order
+        sorted_players = sorted(
+            self.players.values(),
+            key=lambda p: (-p.score, p.name),
+        )
+
+        leaderboard = []
+        current_rank = 0
+        previous_score = None
+
+        for i, player in enumerate(sorted_players):
+            if player.score != previous_score:
+                current_rank = i + 1
+            previous_score = player.score
+
+            entry = {
+                "rank": current_rank,
+                "name": player.name,
+                "score": player.score,
+                "is_admin": player.is_admin,
+                "connected": player.connected,
+                # Final stats (Story 5.6)
+                "best_streak": player.best_streak,
+                "rounds_played": player.rounds_played,
+                "bets_won": player.bets_won,
+            }
+            leaderboard.append(entry)
+
+        return leaderboard
